bot_template/requirements.txt
# Basic Discord Bot Dependencies
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0

# Add other dependencies based on your bot's needs:

# For Notion integration:
# notion-client>=2.0.0

# For AI/OpenAI:
# openai>=1.0.0

# For data processing:
# pandas>=1.5.0
# numpy>=1.21.0

# For image processing:
# Pillow>=8.0.0

# For web scraping:
# beautifulsoup4>=4.10.0
# requests>=2.28.0

# For database:
# sqlalchemy>=1.4.0
# sqlite3 (included in Python)

# For scheduling:
# apscheduler>=3.9.0

# For logging enhancements:
# colorlog>=6.0.0
